Parse examples, build graph, guess query label
Examples use a custom format where one line contains one example, associated
with one set of labels (usually just one label). They are in plain text, with
markup to highlight important words (like `/so/`) and parameters
(like `this [type] [parameter]`.)

We plan on automatically detecting parameters by type, through custom code for
each type. For instance, integers are always of the same form, which a simple
regular expression can match.

The graph is designed so that each vertex corresponds to a word or a parameter,
and each edge signifies that there exists an example in which a particular word
(or parameter) was seen before the word (or parameter) it points to.

Queries to guess are processed by marking each edge walked through when matching
the query to the graph, and summing, for each marked edge, the number of
examples of a particular label that have walked through that edge, normalized
(through division) by the total number of examples for that label. To obtain a
probability, it is then divided by the sum of those numbers for each label.

We guess that the query is of the label with the highest probability.